requirements.txt
beautifulsoup4==4.10.0
Boruta==0.3
catboost==1.0.4
dash==2.0.0
dash_bootstrap_components==1.0.2
Flask_Caching==1.10.1
h2o==3.36.0.3
imbalanced_learn==0.9.0
imblearn==0.0
lightgbm==3.3.2
matplotlib==3.5.1
nba_api==1.1.11
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
plotly==5.5.0
requests==2.27.1
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.1.2
tqdm==4.62.3
vboUtil==0.1.3
verstack==2.0.4
waitress==2.0.0
xgboost==1.3.3
yellowbrick==1.3.post1
